requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
coverage==7.10.4
cssbeautifier==1.15.1
Django==5.2.5
djlint==1.36.1
EditorConfig==0.12.4
et_xmlfile==2.0.0
iniconfig==2.1.0
jsbeautifier==1.15.1
json5==0.9.28
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
six==1.16.0
sqlparse==0.5.1
tqdm==4.67.0
tzdata==2024.2
